% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDmining.R
\docType{package}
\name{IDmining-package}
\alias{IDmining}
\alias{IDmining-package}
\title{Intrinsic Dimension for Data Mining}
\description{
Contains techniques for mining large high-dimensional data sets
by using the concept of Intrinsic Dimension (ID). Here the ID is
not necessarily integer. It is extended to fractal dimensions. And
the Morisita estimator is used for the ID estimation, but other
tools are included as well.
}
\author{
Jean Golay \email{Jean.Golay@unil.ch} and
Mohamed Laib \email{Mohamed.Laib@unil.ch},

Maintainer: Jean Golay \email{Jean.Golay@unil.ch}
}
\references{
J. Golay and M. Kanevski (2015). A new estimator of intrinsic dimension
based on the multipoint Morisita index,
\href{http://www.sciencedirect.com/science/article/pii/S0031320315002320}{Pattern Recognition 48 (12):4070–4081}.

J. Golay, M. Leuenberger and M. Kanevski (2016). Feature Selection for Regression Problems Based
on the Morisita Estimator of Intrinsic Dimension,
\href{https://arxiv.org/abs/1602.00216}{arXiv:1602.00216}.

J. Golay and M. Kanevski (2016). Unsupervised Feature Selection Based on the Morisita Estimator
of Intrinsic Dimension, \href{https://arxiv.org/abs/1608.05581}{arXiv:1608.05581}.

J. Golay, M. Leuenberger and M. Kanevski (2015).
\href{https://www.elen.ucl.ac.be/Proceedings/esann/esannpdf/es2015-41.pdf}{Morisita-based feature selection for regression problems}.
Proceedings of the 23rd European Symposium on Artificial Neural Networks, Computational Intelligence and
Machine Learning (ESANN), Bruges (Belgium).
}

