% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDim.R
\name{RenDim}
\alias{RenDim}
\title{Renyi's Generalized Dimensions}
\usage{
RenDim(X, scaleQ=1:5, qMin=2, qMax=2)
}
\arguments{
\item{X}{A \eqn{N \times E}{N x E} \code{matrix}, \code{data.frame} or \code{data.table} where \eqn{N} is the number
of data points and \eqn{E} is the number of variables (or features). Each variable
is rescaled to the \eqn{[0,1]} interval by the function.}

\item{scaleQ}{A vector (at least two values). It contains the values of \eqn{\ell^{-1}}{l^(-1)}
chosen by the user (by default: \code{scaleQ = 1:5}).}

\item{qMin}{The minimum value of \eqn{q} (by default: \code{qMin = 2}).}

\item{qMax}{The maximum value of \eqn{q} (by default: \code{qMax = 2}).}
}
\value{
A list of two elements:
 \enumerate{
  \item a \code{data.frame} containing the value of Rényi's information of \eqn{qth}{qth} order
  (computed using the natural logarithm) for each value of \eqn{\ln (\delta)}{ln(delta)}
  and \eqn{q}. The values of \eqn{\ln (\delta)}{ln(delta)} are provided with regard to the \eqn{[0,1]} interval.
  \item a \code{data.frame} containing the value of \eqn{D_q}{Dq} for each value of \eqn{q}.
}
}
\description{
Estimates Rényi's generalized dimensions (or Rényi's dimensions of \eqn{qth}{qth} order). It is
mainly for \eqn{q=2}{q=2} that the result is used as an estimate of the intrinsic dimension of data.
}
\details{
\enumerate{
\item \eqn{\ell}{l} is the edge length of the grid cells (or quadrats). Since the variables
(and consenquently the grid) are rescaled to the \eqn{[0,1]} interval, \eqn{\ell}{l} is equal
to \eqn{1} for a grid consisting of only one cell.
\item \eqn{\ell^{-1}}{l^(-1)} is the number of grid cells (or quadrats) along each axis of the
Euclidean space in which the data points are embedded.
\item \eqn{\ell^{-1}}{l^(-1)} is equal to \eqn{Q^{(1/E)}}{Q^(1/E)} where \eqn{Q} is the number
of grid cells and \eqn{E} is the number of variables (or features).
\item \eqn{\ell^{-1}}{l^(-1)} is directly related to \eqn{\delta}{delta} (see References).
\item \eqn{\delta}{delta} is the diagonal length of the grid cells.
}
}
\examples{
sim_dat <- SwissRoll(1000)

scaleQ <- 1:15 # It starts with a grid of 1^E cell (or quadrat).
               # It ends with a grid of 15^E cells (or quadrats).
qRI_ID <- RenDim(sim_dat[,c(1,2)], scaleQ[5:15])

print(paste("The ID estimate is equal to",round(qRI_ID[[1]][1,2],2)))
}
\references{
C. Traina Jr., A. J. M. Traina, L. Wu and C. Faloutsos (2000).
\href{http://repository.cmu.edu/cgi/viewcontent.cgi?article=1572&context=compsci}{Fast feature
selection using fractal dimension}. Proceedings of the 15th Brazilian Symposium
on Databases (SBBD 2000), João Pessoa (Brazil).

E. P. M. De Sousa, C. Traina Jr., A. J. M. Traina, L. Wu and C. Faloutsos (2007). A fast and
effective method to find correlations among attributes in databases,
\href{http://link.springer.com/article/10.1007/s10618-006-0056-4}{Data Mining and
Knowledge Discovery 14(3):367-407}.

J. Golay and M. Kanevski (2015). A new estimator of intrinsic dimension
based on the multipoint Morisita index,
\href{http://www.sciencedirect.com/science/article/pii/S0031320315002320}{Pattern Recognition 48 (12):4070–4081}.

H. Hentschel and I. Procaccia (1983). The infinite number of generalized
dimensions of fractals and strange attractors,
\href{http://www.sciencedirect.com/science/article/pii/016727898390235X}{Physica D 8(3):435-444}.
}
\author{
Jean Golay \email{Jean.Golay@unil.ch}
}
