% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissRoll.R
\name{SwissRoll}
\alias{SwissRoll}
\title{Swiss Roll Data Set Generator}
\usage{
SwissRoll(N=10000)
}
\arguments{
\item{N}{The number of points to be generated.}
}
\value{
A \eqn{N \times 3}{N x 3} data frame containing the
coordinates of the Swiss roll embedded in \eqn{\rm I\!R^3}{R^3}.
}
\description{
Generates random points on the Swiss Roll manifold.
}
\examples{
N <- 1000
sim_dat <- SwissRoll(N)
}
\references{
J. A. Lee and M. Verleysen (2007). Nonlinear Dimensionality Reduction, Springer, New York.
}

