% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDmining.R
\docType{package}
\name{IDmining-package}
\alias{IDmining}
\alias{IDmining-package}
\title{IDmining: Intrinsic Dimension for Data Mining}
\description{
Contains techniques for mining large and high-dimensional data
    sets by using the concept of Intrinsic Dimension (ID). Here the ID is
    not necessarily an integer. It is extended to fractal dimensions. And
    the Morisita estimator is used for the ID estimation, but other
    tools are included as well.
}
\references{
J. Golay and M. Kanevski (2015). A new estimator of intrinsic dimension
based on the multipoint Morisita index,
\href{http://www.sciencedirect.com/science/article/pii/S0031320315002320}{Pattern Recognition 48 (12):4070–4081}.

J. Golay, M. Leuenberger and M. Kanevski (2017). Feature selection for regression problems based
on the Morisita estimator of intrinsic dimension,
\href{http://www.sciencedirect.com/science/article/pii/S0031320317301905}{Pattern Recognition 70:126–138}.

J. Golay and M. Kanevski (2017). Unsupervised feature selection based on the
Morisita estimator of intrinsic dimension,
\href{http://www.sciencedirect.com/science/article/pii/S0950705117303659}{Knowledge-Based Systems 135:125-134}.

J. Golay, M. Leuenberger and M. Kanevski (2015).
\href{https://www.elen.ucl.ac.be/Proceedings/esann/esannpdf/es2015-41.pdf}{Morisita-based feature selection for regression problems}.
Proceedings of the 23rd European Symposium on Artificial Neural Networks, Computational Intelligence and
Machine Learning (ESANN), Bruges (Belgium).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.sites.google.com/site/jeangolayresearch/}
}

}
\author{
Jean Golay \email{jeangolay@gmail.com} and
Mohamed Laib \email{laib.med@gmail.com},

Maintainer: Jean Golay \email{jeangolay@gmail.com}
}
